Implemented user CRUD

+++ services/implementations/UserService.ts
@@ -1,14 +1,16 @@
-import { User, VideoState } from "@prisma/client";
-import { IVideoRepository } from "../../repositories/IVideoRepository";
-import { IVideoService } from "../IVideoService";
+import { User } from "@prisma/client";
 import { NotFoundError } from "../../errors/NotFoundError";
-import { IVideoFileService } from "../IVideoFileService";
 import { IUserService } from "../IUserService";
 import { IUserRepository } from "../../repositories/IUserRepository";
 
 export class UserService implements IUserService {
     constructor(private userRepository: IUserRepository) {
 
+    }
+    async authenticateUser(email: string, password: string): Promise<{ id: string; email: string; password: string; displayName: string | null; }> {
+        let user = await this.userRepository.authenticateUser(email, password);
+        if (user == null) throw new NotFoundError(401, "Invalid email or password");
+        return user;
     }
     async getAllUsers(): Promise<User[]> {
         return await this.userRepository.findAllUsers();
@@ -19,7 +21,11 @@ export class UserService implements IUserService {
         return video;
     }
     async deleteUserById(id: string): Promise<User> {
-        return await this.userRepository.deleteUserById(id);
+        let user = await this.userRepository.findUserById(id);
+        if (!user) throw new NotFoundError(404, "User not found");
+        else {
+            return await this.userRepository.deleteUserById(id);
+        }
     }
     async createUser(email: string, password: string, displayName: string): Promise<User> {
         return await this.userRepository.createUser(email, password, displayName);
